Add navigation menu component
This commit adds a new component called NavMenu, which is responsible for rendering a navigation menu with links. The NavMenu component accepts a className prop for styling customization. It uses an array of links to generate the menu structure, including support for nested submenus. The links are defined in the NavMenu.tsx file, and the menu is rendered in the layout.tsx file.
