+++ app/layout.tsx
@@ -1,45 +1,9 @@
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import "./globals.css";
+import { NavMenu } from "./NavMenu";
 
 const inter = Inter({ subsets: ["latin"] });
-
-const links = [
-  {
-    name: "X",
-    url: "https://x.com/thebrianalonso",
-  },
-  {
-    name: "Blog",
-    url: "https://blog.balonso.com",
-  },
-  {
-    name: "Projects",
-    links: [
-      {
-        name: "Random Footballer",
-        url: "https://randomfootballer.com",
-      },
-      {
-        name: "Do Not Waste Your Life",
-        url: "https://donotwasteyourlife.com",
-      },
-    ],
-  },
-  {
-    name: "Products",
-    links: [
-      {
-        name: "LuckyLink",
-        url: "https://luckylink.app",
-      },
-      {
-        name: "EarthWallet",
-        url: "https://earthwallet.app",
-      },
-    ],
-  },
-];
 
 export const metadata: Metadata = {
   title: "Brian Alonso",
@@ -83,30 +47,7 @@ export default function RootLayout({
             </a>
           </div>
           <div className="navbar-center hidden md:flex">
-            <ul className="menu menu-horizontal px-1">
-              {links.map((link) => (
-                <li key={link.name}>
-                  {link.links ? (
-                    <details>
-                      <summary>{link.name}</summary>
-                      <ul>
-                        {link.links.map((sublink) => (
-                          <li key={sublink.name}>
-                            <a href={sublink.url} target="_blank">
-                              {sublink.name}
-                            </a>
-                          </li>
-                        ))}
-                      </ul>
-                    </details>
-                  ) : (
-                    <a href={link.url} target="_blank">
-                      {link.name}
-                    </a>
-                  )}
-                </li>
-              ))}
-            </ul>
+            <NavMenu className="menu menu-horizontal px-1" />
           </div>
           <div className="flex-shrink md:navbar-end">
             <label className="swap swap-rotate">
@@ -158,30 +99,7 @@ export default function RootLayout({
                         />
                       </svg>
                     </summary>
-                    <ul className="dropdown-content mt-3 z-[1] p-2 shadow bg-base-100 rounded-box absolute right-0">
-                      {links.map((link) => (
-                        <li key={link.name}>
-                          {link.links ? (
-                            <details>
-                              <summary>{link.name}</summary>
-                              <ul>
-                                {link.links.map((sublink) => (
-                                  <li key={sublink.name}>
-                                    <a href={sublink.url} target="_blank">
-                                      {sublink.name}
-                                    </a>
-                                  </li>
-                                ))}
-                              </ul>
-                            </details>
-                          ) : (
-                            <a href={link.url} target="_blank">
-                              {link.name}
-                            </a>
-                          )}
-                        </li>
-                      ))}
-                    </ul>
+                    <NavMenu className="dropdown-content mt-3 z-[1] p-2 shadow bg-neutral text-neutral-content rounded-box absolute right-0" />
                   </details>
                 </li>
               </ul>
